update developer guide
